requirements.txt
Django
djangorestframework
drf-spectacular
importlib_metadata
psycopg2-binary
gunicorn
whitenoise
django-tailwind
django-browser-reload
markdown
django-filter